refactor(topology): split flowYamlUtils into one module per concern
One 1175-line file did three unrelated jobs: plain-object serialization
through js-yaml, mutation of the yaml AST, and read-only cursor queries
for the editor. Mixing read-only position lookups with AST surgery in one
file is why "is it safe to touch this" was a hard question.

Now seven modules under src/utils/yaml/:

  document.ts       27  parseDocumentTyped, scalarKey, TOSTRING_OPTIONS
  paths.ts          73  dotted paths to and from segment arrays, no yaml
  serialization.ts 123  parse, stringify, pairsToMap
  fields.ts        103  every map with a given field, every typed block
  metadata.ts      153  the flow's root keys
  blocks.ts        453  extract, insert, replace, delete, swap
  positions.ts     301  what sits at a cursor or on a line

flowYamlUtils.ts is now an 11-line barrel, so the
`@kestra-io/topology/flow-yaml-utils` subpath and every `import * as`
consumer are untouched. Verified rather than assumed: the runtime export
surface is identical, all 29 names, diffed against the parent commit.

Dependencies run one way — paths and serialization import nothing local,
document imports nothing local, fields and metadata use document, blocks
uses document and paths, positions uses document and fields.

The 1632-line test file is split the same way, into six files that each
import the module they exercise rather than the barrel, so a test fails
loudly if the export it covers moves. Two block tests use parse and
stringify as fixtures and import them from serialization directly.

Noticed while splitting, left alone: root keys get ordered twice by two
different lists, SORT_FIELDS in serialization (applied by
replaceBlockWithPath) and ORDERED_FLOW_ROOT_KEYS in metadata (applied by
updateMetadata). They disagree, and reconciling them changes output.
